requirements.txt
numpy==1.18.5
opencv-python==4.4.0.42
protobuf==3.12.4
pyzmq==19.0.2
tensorflow==2.3.0
